extract util components out of navbar
